requirements.txt
mcp>=1.25.0
httpx>=0.28.0
python-dotenv>=1.0.0
